Naming adjustments, simplify NftMetadataService
